pyproject.toml
[tool.poetry]
name = "getxerpa"
version = "0.1.0"
description = ""
authors = ["Saúl Hormazábal <saul.hormazabal@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
Django = "^5.0"
djangorestframework = "^3.14.0"
python-decouple = "^3.8"
psycopg2-binary = "^2.9.9"
pillow = "^10.1.0"
drf-yasg = "^1.21.7"
setuptools = "^69.0.3"
django-pgviews-redux = "^0.9.4"

[tool.poetry.group.dev.dependencies]
bpython = "^0.24"
pdbpp = "^0.10.3"
django-debug-toolbar = "^4.2.0"
django-extensions = "^3.2.3"
werkzeug = "^3.0.1"
coverage = "^7.4.0"
dodgy = "^0.2.1"
flake8 = "^6.1.0"
pylint = "^3.0.3"
pylint-django = "^2.5.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
